# artifacts/openapi.json
{
  "components": {
    "schemas": {
...
      "CalendarConnectionSummary": {
        "properties": {
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "mode",
          "status"
        ],
        "title": "CalendarConnectionSummary",
        "type": "object"
      },
      "CalendarEventSummary": {
        "properties": {
          "all_day": {
            "title": "All Day",
            "type": "boolean"
          },
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "sync_status": {
            "title": "Sync Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "start_at",
          "end_at",
          "all_day",
          "kind",
          "sync_status"
        ],
        "title": "CalendarEventSummary",
        "type": "object"
      },
      "CalendarResponse": {
        "properties": {
          "connections": {
            "items": {
              "$ref": "#/components/schemas/CalendarConnectionSummary"
            },
            "title": "Connections",
            "type": "array"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/CalendarEventSummary"
            },
            "title": "Events",
            "type": "array"
          },
          "flexible_tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSummary"
            },
            "title": "Flexible Tasks",
            "type": "array"
          }
        },
        "required": [
          "connections",
          "events",
          "flexible_tasks"
        ],
        "title": "CalendarResponse",
        "type": "object"
      },
...
      "ExternalDependencySummary": {
        "properties": {
          "deliverable": {
            "title": "Deliverable",
            "type": "string"
          },
          "expected_delivery_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Delivery At"
          },
          "fallback_available": {
            "title": "Fallback Available",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "owner_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Label"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "deliverable",
          "owner_label",
          "expected_delivery_at",
          "fallback_available",
          "version"
        ],
        "title": "ExternalDependencySummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InboxCandidate": {
        "properties": {
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "evidence_score": {
            "title": "Evidence Score",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inbox_item_id": {
            "title": "Inbox Item Id",
            "type": "string"
          },
          "interpretation": {
            "additionalProperties": true,
            "title": "Interpretation",
            "type": "object"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "source_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Text"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "inbox_item_id",
          "kind",
          "status",
          "source_text",
          "interpretation",
          "evidence_score",
          "decision"
        ],
        "title": "InboxCandidate",
        "type": "object"
      },
      "InboxResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/InboxCandidate"
            },
            "title": "Candidates",
            "type": "array"
          }
        },
        "required": [
          "candidates"
        ],
        "title": "InboxResponse",
        "type": "object"
      },
...
      "MilestoneSummary": {
        "properties": {
          "deadline_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "deadline_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Date"
          },
          "deadline_time_known": {
            "title": "Deadline Time Known",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "deadline_date",
          "deadline_at",
          "deadline_time_known",
          "version"
        ],
        "title": "MilestoneSummary",
        "type": "object"
      },
...
      "ProjectDetailResponse": {
        "properties": {
          "external_dependencies": {
            "items": {
              "$ref": "#/components/schemas/ExternalDependencySummary"
            },
            "title": "External Dependencies",
            "type": "array"
          },
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneSummary"
            },
            "title": "Milestones",
            "type": "array"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectSummary"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSummary"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "project",
          "milestones",
          "tasks",
          "external_dependencies"
        ],
        "title": "ProjectDetailResponse",
        "type": "object"
      },
      "ProjectSummary": {
        "properties": {
          "done_count": {
            "title": "Done Count",
            "type": "integer"
          },
          "execution_progress": {
            "title": "Execution Progress",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "risk_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Risk Reasons",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_count": {
            "title": "Task Count",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "execution_progress",
          "risk_level",
          "risk_reasons",
          "task_count",
          "done_count"
        ],
        "title": "ProjectSummary",
        "type": "object"
      },
      "ProjectsResponse": {
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            },
            "title": "Projects",
            "type": "array"
          }
        },
        "required": [
          "projects"
        ],
        "title": "ProjectsResponse",
        "type": "object"
      },
...
      "ProposalSummary": {
        "properties": {
          "approval_level": {
            "title": "Approval Level",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "targets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Targets",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "approval_level",
          "status",
          "version",
          "targets"
        ],
        "title": "ProposalSummary",
        "type": "object"
      },
      "ReviewResponse": {
        "properties": {
          "actual_minutes": {
            "title": "Actual Minutes",
            "type": "integer"
          },
          "missed_minutes": {
            "title": "Missed Minutes",
            "type": "integer"
          },
          "pending_proposals": {
            "items": {
              "$ref": "#/components/schemas/ProposalSummary"
            },
            "title": "Pending Proposals",
            "type": "array"
          },
          "period_end": {
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date",
            "title": "Period Start",
            "type": "string"
          },
          "planned_minutes": {
            "title": "Planned Minutes",
            "type": "integer"
          }
        },
        "required": [
          "period_start",
          "period_end",
          "planned_minutes",
          "actual_minutes",
          "missed_minutes",
          "pending_proposals"
        ],
        "title": "ReviewResponse",
        "type": "object"
      },
      "SessionResponse": {
        "properties": {
          "token": {
            "title": "Token",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "token",
          "user_id"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "TaskSummary": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "remaining_minutes": {
            "title": "Remaining Minutes",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "remaining_minutes",
          "version"
        ],
        "title": "TaskSummary",
        "type": "object"
      },
...
      "TestSessionRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "TestSessionRequest",
        "type": "object"
      },
      "TodayResponse": {
        "properties": {
          "core_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "fixed_events": {
            "items": {
              "$ref": "#/components/schemas/FixedEvent"
            },
            "title": "Fixed Events",
            "type": "array"
          },
          "must_do": {
            "items": {
              "$ref": "#/components/schemas/TaskSummary"
            },
            "title": "Must Do",
            "type": "array"
          },
          "not_today": {
            "items": {
              "$ref": "#/components/schemas/TaskSummary"
            },
            "title": "Not Today",
            "type": "array"
          },
          "plan_status": {
            "title": "Plan Status",
            "type": "string"
          },
          "queue": {
            "items": {
              "$ref": "#/components/schemas/TaskSummary"
            },
            "title": "Queue",
            "type": "array"
          }
        },
        "required": [
          "plan_status",
          "core_outcome",
          "fixed_events",
          "must_do",
          "queue",
          "not_today"
        ],
        "title": "TodayResponse",
        "type": "object"
      },
...
      "UploadInitiationResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "storage_key": {
            "title": "Storage Key",
            "type": "string"
          },
          "upload_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Url"
          }
        },
        "required": [
          "id",
          "storage_key",
          "status"
        ],
        "title": "UploadInitiationResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Personal PM Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/calendar": {
      "get": {
        "operationId": "get_calendar_api_v1_calendar_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Calendar",
        "tags": [
          "views"
        ]
      }
    },
...
    "/api/v1/identity/test-session": {
      "post": {
        "operationId": "create_test_session_api_v1_identity_test_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Test Session",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/inbox": {
      "get": {
        "operationId": "get_inbox_api_v1_inbox_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Inbox",
        "tags": [
          "views"
        ]
      }
    },
    "/api/v1/inbox/sources": {
      "post": {
        "operationId": "upload_source_api_v1_inbox_sources_post",
        "parameters": [
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Filename",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadInitiationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Source",
        "tags": [
          "source-artifacts"
        ]
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "operationId": "get_projects_api_v1_projects_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Projects",
        "tags": [
          "views"
        ]
      }
    },
    "/api/v1/projects/{workstream_id}": {
      "get": {
        "operationId": "get_project_detail_api_v1_projects__workstream_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workstream_id",
            "required": true,
            "schema": {
              "title": "Workstream Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Detail",
        "tags": [
          "views"
        ]
      }
    },
...
    "/api/v1/review": {
      "get": {
        "operationId": "get_review_api_v1_review_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Review",
        "tags": [
          "views"
        ]
      }
    },
...
    "/api/v1/today": {
      "get": {
        "operationId": "get_today_api_v1_today_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Today",
        "tags": [
          "views"
        ]
      }
    },
...
  }
}